refactor: import tycho pb structs from tycho-protocol-sdk
