+++ Cargo.toml
@@ -24,7 +24,7 @@ debug = 0
 [workspace.dependencies]
 chrono = { version = "0.4.26", features = ["serde"] }
 serde_json = "1.0.105"
-serde = { version = "1.0", features = ["rc"] }
+serde = { version = "1.0.204", features = ["rc"] }
 strum = "0.25.0"
 strum_macros = "0.25.2"
 alloy-primitives = { version = "0.6.2", features = ["serde"] }
