+++ tycho-indexer/Cargo.toml
@@ -64,6 +64,7 @@ mini-moka = "0.10.3"
 num-bigint = "0.4.4"
 num-traits = "0.2.19"
 num_cpus = "1.16.0"
+tycho-substreams = { git = "https://github.com/propeller-heads/tycho-protocol-sdk.git", rev = "04d14db" }
 
 [dev-dependencies]
 actix-rt = "2.9.0"
