+++ tycho-indexer/src/pb/mod.rs
@@ -2,8 +2,6 @@ use std::fmt::Display;
 
 use crate::pb::sf::substreams::rpc::v2::BlockRange;
 
-pub mod testing;
-
 include!("pb.rs");
 
 impl Display for BlockRange {
